added user id in mozark.eventAtrribute

+++ src/main/java/in/mozark/service/impl/EventServiceImpl.java
@@ -7,10 +7,7 @@ import java.io.FileOutputStream;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.lang.reflect.Field;
-import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -25,6 +22,7 @@ import org.springframework.util.CollectionUtils;
 import org.springframework.util.StringUtils;
 
 import in.mozark.dto.EventAttributeDto;
+import in.mozark.dto.MozarkEventAttributeDto;
 import in.mozark.dto.ResponseDto;
 import in.mozark.dto.UploadFileDto;
 import in.mozark.service.EventService;
@@ -61,7 +59,7 @@ public class EventServiceImpl implements EventService {
 				logger.info("fileNameAndFileType {} ", fileNameAndFileType);
 
 				Map<String, ResponseDto> uploadResponse = connectUploadAPI(files, fileNameAndFileType,
-						uploadFileDto.getEventName(), eventAttributeDto.getTestId());
+						uploadFileDto);
 			  logger.info("uploadResponse {} ",uploadResponse);
 			  response.setResponseCode("00");
 			  response.setResponseMessage("sucess");
@@ -81,7 +79,7 @@ public class EventServiceImpl implements EventService {
 	}
 
 	private Map<String, ResponseDto> connectUploadAPI(File[] files, Map<String, String> fileNameAndFileType,
-			String eventName, String testId) {
+			UploadFileDto inputUploadFileDto) {
 		logger.info("fileNameAndFileType {} ", fileNameAndFileType);
 		ResponseDto response = new ResponseDto();
 		Map<String, ResponseDto> uploadAndDeletResponse = new HashMap<String, ResponseDto>();
@@ -89,39 +87,65 @@ public class EventServiceImpl implements EventService {
 		for (File file : files) {
 
 			try {
+				
+				String eventName = inputUploadFileDto.getEventName();
+				logger.info("eventName - {} ",eventName);
+				
+				EventAttributeDto inputEventAttributeDto = inputUploadFileDto.getEventAttributeDto();
+				logger.info("eventAttributeDto {} ", inputEventAttributeDto);
+				
+				String testId = inputEventAttributeDto.getTestId();
+				logger.info("testId - {} ",testId);
+				
+				MozarkEventAttributeDto inputMozarkAttributeDto = inputUploadFileDto.getMozarkEventAttributeDto();
+				logger.info("inputMozarkAttributeDto - {} ",inputMozarkAttributeDto);
+				
 				fileName = file.getName();
 				logger.info("fileName {} ", fileName);
 
 				String fileType = fileNameAndFileType.get(fileName);
 				logger.info("fileType {} ", fileType);
-				UploadFileDto uploadFileDto = new UploadFileDto();
-				uploadFileDto.setEventName(eventName);
-				uploadFileDto.setFileData(file);
+				
+				if(!StringUtils.isEmpty(fileType)) {
+					UploadFileDto uploadFileDto = new UploadFileDto();
+					uploadFileDto.setEventName(eventName);
+					uploadFileDto.setFileData(file);
 
-				EventAttributeDto eventAttributeDto = new EventAttributeDto();
+					EventAttributeDto eventAttributeDto = new EventAttributeDto();
 
-				eventAttributeDto.setFileName(fileName);
-				eventAttributeDto.setFileType(fileType);
-				eventAttributeDto.setTestId(testId);
-				uploadFileDto.setEventAttributeDto(eventAttributeDto);
+					eventAttributeDto.setFileName(fileName);
+					eventAttributeDto.setFileType(fileType);
+					eventAttributeDto.setTestId(testId);
+					uploadFileDto.setEventAttributeDto(eventAttributeDto);
+					
+					MozarkEventAttributeDto mozarkEventAttributeDto = new MozarkEventAttributeDto();
+					mozarkEventAttributeDto.setUserId(inputMozarkAttributeDto.getUserId());
+					uploadFileDto.setMozarkEventAttributeDto(mozarkEventAttributeDto);
 
-				logger.info("uploadFileDto {} ", uploadFileDto.toString());
-				ResponseDto uploadAPIResponse = connectingMozarkAPI.connectingUploadAPI(uploadFileDto);
-				logger.info("uploadAPIResponse {} ", uploadAPIResponse);
+					logger.info("uploadFileDto {} ", uploadFileDto.toString());
+					ResponseDto uploadAPIResponse = connectingMozarkAPI.connectingUploadAPI(uploadFileDto);
+					logger.info("uploadAPIResponse {} ", uploadAPIResponse);
 
-				if (uploadAPIResponse.getResponseCode().equalsIgnoreCase("00")) {
-					File fileToBeDeleted = new File(fileUploadDirectory + testId + "/" + fileName);
-					if (fileToBeDeleted.delete()) {
-						logger.info("file deleted successfully ");
-						response.setResponseCode("00");
-						response.setResponseMessage("sucess");
-					} else {
-						logger.info("file didn't deleted successfully");
-						response.setResponseCode("01");
-						response.setResponseMessage("failed");
+					if (uploadAPIResponse.getResponseCode().equalsIgnoreCase("00")) {
+						File fileToBeDeleted = new File(fileUploadDirectory + testId + "/" + fileName);
+						if (fileToBeDeleted.delete()) {
+							logger.info("file deleted successfully ");
+							response.setResponseCode("00");
+							response.setResponseMessage("sucess");
+						} else {
+							logger.info("file didn't deleted successfully");
+							response.setResponseCode("01");
+							response.setResponseMessage("failed");
+						}
+						
 					}
-					uploadAndDeletResponse.put(fileName, response);
+						
+				}else {
+					response.setResponseCode("99");
+					response.setResponseMessage("File Type is not available");
+					
 				}
+				uploadAndDeletResponse.put(fileName, response);
 
 			} catch (Exception e) {
 				logger.error("Exception occurred in connectUploadAPI {} ", e.getMessage());
@@ -142,30 +166,32 @@ public class EventServiceImpl implements EventService {
 		EventAttributeDto eventAttributeDto  = uploadFileDto.getEventAttributeDto();
 		logger.info("eventAttributeDto {} ",eventAttributeDto);
 		
-		String command = eventAttributeDto.getCommand();
-		logger.info("command {} ",command);
+		
 		try {
-			Process process =Runtime.getRuntime().exec(command);
 			
-			long pid = -1;
-            try {
-              if (process.getClass().getName().equals("java.lang.UNIXProcess")) {
-                Field f = process.getClass().getDeclaredField("pid");
-                f.setAccessible(true);
-                pid = f.getLong(process);
-                f.setAccessible(false);
-              }
-            } catch (Exception e) {
-            	logger.error("exception occurred while getting process id - ",e.getMessage());
-              pid = -1;
-            }
-            
-	        String processId = String.valueOf(pid);
-	        logger.info("processId {} ",processId);
 	        
 	       String action =  eventAttributeDto.getAction();
 	       logger.info("action {} ",action);
 	       if(!StringUtils.isEmpty(action) && action.equalsIgnoreCase("start")) {
+	    	   String command = eventAttributeDto.getCommand();
+	   		   logger.info("command {} ",command);
+	    	   Process process =Runtime.getRuntime().exec(command);
+				
+				long pid = -1;
+	            try {
+	              if (process.getClass().getName().equals("java.lang.UNIXProcess")) {
+	                Field f = process.getClass().getDeclaredField("pid");
+	                f.setAccessible(true);
+	                pid = f.getLong(process);
+	                f.setAccessible(false);
+	              }
+	            } catch (Exception e) {
+	            	logger.error("exception occurred while getting process id - ",e.getMessage());
+	              pid = -1;
+	            }
+	            
+		        String processId = String.valueOf(pid);
+		        logger.info("processId {} ",processId);
 	    	   File folderCreate = new File(fileUploadDirectory+eventAttributeDto.getTestId()+"/");
 				if (!folderCreate.exists()) {
 					if (folderCreate.mkdirs()) {
@@ -221,9 +247,11 @@ public class EventServiceImpl implements EventService {
 								if (System.getProperty("os.name").toLowerCase().indexOf("windows") > -1) {
 									 Process p = rt.exec("taskkill " +id);
 									 p.destroy();
+									 logger.info("killed process id - ",id);
 								}else {
 									Process p = rt.exec("kill -9 " +id);
 									p.destroy();
+									logger.info("killed process id - ",id);
 								}
 							    
 							} catch (IOException e) {
